Fixed last commit: added new functionality for 64-bits auxiliary analog I/O ports on main struct of irioCore

+++ src/main/c/include/irioDataTypes.h
@@ -185,7 +185,7 @@ typedef struct TResourcePort{
  */
 typedef struct TResourcePort_64{
 	uint8_t found;
-	uint32_t value;
+	uint64_t value;
 }TResourcePort_64;
 
 /**
